+++ libshviotqt/src/rpc/tcpserver.cpp
@@ -23,7 +23,7 @@ TcpServer::~TcpServer()
 bool TcpServer::start(int port)
 {
 	shvInfo() << "start listenning on port:" << port;
-	if (!listen(QHostAddress::AnyIPv4, port)) {
+	if (!listen(QHostAddress::AnyIPv4, static_cast<quint16>(port))) {
 		shvError() << tr("Unable to start the server: %1.").arg(errorString());
 		close();
 		return false;
